Collison code
finished collison code which will detect a collison between an array of objects. Please test for bugs and if you find any I will try to fix. The collison code uses getbound of object method to determine if points of different objects are in a one pixel range of each other, if so one of those points will be pushed to an array that will be returned when all points have been checked against all other points.